+++ elements/src/components.d.ts
@@ -100,6 +100,10 @@ export interface PosAddLiteralValueCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLPosAddLiteralValueElement;
 }
+export interface PosAppCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLPosAppElement;
+}
 export interface PosAppDocumentViewerCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLPosAppDocumentViewerElement;
@@ -219,7 +223,18 @@ declare global {
         prototype: HTMLPosAddNewThingElement;
         new (): HTMLPosAddNewThingElement;
     };
+    interface HTMLPosAppElementEventMap {
+        "pod-os:session-restored": { url: string };
+    }
     interface HTMLPosAppElement extends Components.PosApp, HTMLStencilElement {
+        addEventListener<K extends keyof HTMLPosAppElementEventMap>(type: K, listener: (this: HTMLPosAppElement, ev: PosAppCustomEvent<HTMLPosAppElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLPosAppElementEventMap>(type: K, listener: (this: HTMLPosAppElement, ev: PosAppCustomEvent<HTMLPosAppElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
     }
     var HTMLPosAppElement: {
         prototype: HTMLPosAppElement;
@@ -722,6 +737,10 @@ declare namespace LocalJSX {
         "referenceUri": string;
     }
     interface PosApp {
+        /**
+          * Fired whenever the session was restored
+         */
+        "onPod-os:session-restored"?: (event: PosAppCustomEvent<{ url: string }>) => void;
         "restorePreviousSession"?: boolean;
     }
     interface PosAppBrowser {